preparing plant-simulate

+++ tools/simulate.py
@@ -1,41 +1,11 @@
 from models.plant import Plant
 from models.sensor import SensorData, SensorDataPrediction
+from tools.sensor import ToolChainSensor
 from tools.forecast import SensorDataForecast
 from mesh_network.dedicated import MeshDedicatedDispatch
 import random
 import datetime
 
-
-# class SimulatePlant():
-#   def __init__(self, plant, sensor):
-#     self.plant = plant
-#     self.sensor = sensor
-
-#     data = SensorData.select().where(SensorData.plant == plant,
-#                                      SensorData.sensor == sensor)
-
-#     if data.count() < 1:
-#       self.__copy(sensor)
-
-#   def __copy(self, sensor):
-#     plant_sensor_data = {}
-#     for plant in Plant.select():
-#       data = SensorData.select().where(SensorData.plant == plant,
-#                                        SensorData.sensor == sensor)
-
-#       plant_sensor_data[plant.name] = data.count
-
-#     largest = sorted(plant_sensor_data)[0]
-
-#     # simulated plant -> copy to plant and then proceed
-#     # new timestamps! c:
-
-#   def run(self):
-#     # see if in table -> create new? or use old?
-#     #  -> if 0 -> run regression again!
-#     # grab earliest -> put into database and delete from table
-#     # call MeshNetworkDispatch
-#     pass
 
 class PlantSimulate:
   def __init__(self):
@@ -96,16 +66,12 @@ class PlantSimulate:
                                              .limit(1)
 
     for sample in predicted:
-      print(type(sample.time))
-      new = SensorData()
-      new.plant = target
-      new.sensor = sensor
-      new.value = predicted.value
-      new.created_at = sample.time
-      new.save()
-
-      sample.delete_instance()
-
-    MeshDedicatedDispatch().new_data(sensor)
-
-    return True
+      if sample.time >= datetime.datetime.now():
+        releaser = True
+      	ToolChainSensor().insert_sensor({'sensor': sensor, 'plant': target, 'value': sample.value})
+        # insert hardware? -- test if DUMMYPLANT hardware?
+        sample.delete_instance()
+      else:
+        released = False
+        print('not in prediction timeframe')
+    return released
